+++ src/libMinMax/MinMax.h
@@ -2,5 +2,5 @@
 
 class MinMax {
 public:
-    static Board minmax(const Board& board, bool turn, int depth, bool start);
+    static Board minmax(const Board& board, bool turn, double depth, bool start);
 };
